fixed bug when phi=0
[gitreformat yapf-ify (github/ghtdak) on Fri Dec  4 04:49:17 2015]
[from commit: e265f441184d1dafa0d7c013092bf25aff516837]

+++ lib/common.py
@@ -410,7 +410,7 @@ def weighted_order_choose(orders, n, feekey):
     if phi > 0:
         weight = np.exp(-(1.0 * fee - minfee) / phi)
     else:
-        weight = np.ones_like(fee)
+        weight = np.ones_like(fee, dtype=np.float)
     weight /= sum(weight)
     debug('randomly choosing orders with weighting\n' + pprint.pformat(zip(
         orders, weight)))
